+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+## 2025-02-25
+### v1.0.2
+- Add import file and folder function
+- Add export file function
+
+
 ## 2025-02-24
 ### v1.0.1
 - Add undo and redo function